Add option different chrome layouts.
