+++ src/js/App/RootApp/RootApp.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import { isBeta } from '../../utils';
+import ScalprumRoot from './ScalprumRoot';
+
+const RootApp = (props) => (
+  <BrowserRouter basename={isBeta() ? '/beta' : '/'}>
+    <ScalprumRoot {...props} />
+  </BrowserRouter>
+);
+
+export default RootApp;
