+++ src/js/App/Sidenav/Navigation/ChromeLink.js
@@ -121,6 +121,7 @@ RefreshLink.propTypes = {
   isExternal: PropTypes.bool,
   onLinkClick: PropTypes.func,
   isBeta: PropTypes.bool,
+  currAppId: PropTypes.any,
 };
 
 const ChromeLink = ({ appId, children, ...rest }) => {
